video_classifier/requirements.txt
transformers
torchvision
fastapi
python-multipart
uvicorn[standard]
PyJWT
google-auth
dlib>=19.24.2
matplotlib==3.8.3
numpy==1.26.4
opencv_python==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
retina_face==0.0.14
seaborn==0.13.2
torch
torchvision
tqdm==4.66.1
urllib3==2.2.1
tf-keras==2.19.0